Interface implementada para comunicação Adapter <-> View<-> VM / Inicialização do BD corrigida

+++ app/src/main/java/com/example/movieshearch/service/repository/FavoriteMovieRepository.kt
@@ -1,36 +1,46 @@
 package com.example.movieshearch.service.repository
 
-import android.content.Context
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import com.example.movieshearch.MyApplication
 import com.example.movieshearch.model.MovieModel
-import com.example.movieshearch.service.repository.local.MovieDataBase
 
-class FavoriteMovieRepository(context: Context){
+class FavoriteMovieRepository() {
 
     //liveData
     private val _movieListData = MutableLiveData<MutableList<MovieModel>>()
     val movieListData: LiveData<MutableList<MovieModel>>
-    get() = _movieListData
+        get() = _movieListData
 
     //access BD with DAO
-    private val mDataBase = MovieDataBase.getInstance(context).movieDAO()
+    private val mDataBase = MyApplication.database?.movieDAO()
+    private var saved: Boolean = false
+     var obtainedById: MovieModel = MovieModel()
 
     fun save(movie: MovieModel): Boolean {
-        return mDataBase.save(movie) > 0
+        if (mDataBase != null) {
+            saved = mDataBase.save(movie) > 0
+        }
+        return saved
     }
 
     fun getAll(): MutableList<MovieModel> {
-        _movieListData.value = mDataBase.getAll()
+        if (mDataBase != null) {
+            _movieListData.value = mDataBase.getAll()
+        }
         return movieListData.value!!
     }
 
     fun getById(id: String): MovieModel {
-        return mDataBase.get(id)
+        if (mDataBase != null) {
+            obtainedById = mDataBase.get(id)
+        }
+        return obtainedById
     }
 
     fun remove(movie: MovieModel) {
-        return mDataBase.remove(movie)
+        if (mDataBase != null) {
+            return mDataBase.remove(movie)
+        }
     }
-
 }
